Add check if modmail was send via DM and only delete invokation if not

+++ bot/moderation/moderation.py
@@ -771,7 +771,7 @@ class ModerationCog(commands.Cog):
         Args:
             ctx (discord.ext.commands.Context): The context in which the command was called.
         """
-        if not self._db_connector.is_botonly(ctx.channel.id):
+        if ctx.channel.type.name not in ["private", "group"] and not self._db_connector.is_botonly(ctx.channel.id):
             await ctx.message.delete()
 
         msg_content = ctx.message.content[len(ctx.prefix + ctx.command.name):]
